Rename encoding module to internal
