+++ internal/encoding_test.go
@@ -1,4 +1,4 @@
-package encoding
+package internal
 
 import (
 	"testing"
